+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Changed HTML4.0 and HTML5 renderer to embed images as data URLs
  * Fixed HTML5 renderer image scaling with javascript disabled
  * Added workaround for aborting background worker thread
+ * Added missing ImageRenderer resources
 
 # 15.1.11
  * Added remote report support